Add Search to Party stash
